+++ modules/apps/document-library/document-library-preview-pdf/src/main/java/com/liferay/document/library/preview/pdf/internal/portlet/action/EditPDFPreviewMVCActionCommand.java
@@ -69,7 +69,7 @@ public class EditPDFPreviewMVCActionCommand extends BaseMVCActionCommand {
 
 		try {
 			_pdfPreviewManagedServiceFactory.updatePDFPreview(
-				ParamUtil.getLong(actionRequest, "maxNumberOfPages"), scope,
+				ParamUtil.getInteger(actionRequest, "maxNumberOfPages"), scope,
 				scopePK);
 		}
 		catch (ConfigurationModelListenerException | PDFPreviewException
